requirements.txt
# =============================================================================
# MAFA-Agents Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.24.0

# -----------------------------------------------------------------------------
# AI & LLM Framework
# -----------------------------------------------------------------------------
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
google-generativeai>=0.3.0

# -----------------------------------------------------------------------------
# Model Context Protocol (MCP)
# -----------------------------------------------------------------------------
mcp>=0.2.0
langchain-mcp-adapters>=0.1.0

# -----------------------------------------------------------------------------
# Event Bus & Async
# -----------------------------------------------------------------------------
redis[asyncio]>=5.0.0
websockets>=11.0

# -----------------------------------------------------------------------------
# Machine Learning & Data
# -----------------------------------------------------------------------------
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow-cpu>=2.15.0
joblib>=1.3.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
supabase>=2.0.0
requests>=2.31.0

# -----------------------------------------------------------------------------
# Data Fallback (Yahoo Finance)
# -----------------------------------------------------------------------------
yfinance>=0.2.36
